Fix QG
